Refactor sigmoid functions to separate file
