+++ simple_neuralnet_python/__init__.py
@@ -1,3 +1,4 @@
-from .neuralnet import NeuralNet
+from simple_neuralnet_python.neuralnet import NeuralNet
+from simple_neuralnet_python.sigmoid import sigmoid, sigmoid_derivative
 
-__all__ = ['NeuralNet']
+__all__ = ['NeuralNet', 'sigmoid', 'sigmoid_derivative']
